requirements.txt
opencv-python
mediapipe
moderngl
moderngl-window
numpy
pyrr
pywavefront
Pillow
